Separate JSX into components
