+++ src/components/index.ts
@@ -1,2 +1,7 @@
-export * from './Course';
-export * from './Contact';
+export { Course } from './Course';
+export { Contact } from './Contact';
+export { FloatingActionable } from './FloatingActionable';
+export { MobileNavigation } from './MobileNavigation';
+export { Navigation } from './Navigation';
+export { TagBar } from './TagBar';
+export { TopBar } from './TopBar';
