+++ src/config/index.ts
@@ -1,2 +1,2 @@
-export * from './firebase';
-export * from './formSubmit';
+export { firebaseConfig } from './firebase';
+export { formSubmit } from './formSubmit';
