+++ src/utils/index.ts
@@ -1,2 +1,2 @@
-export * from './getWidth';
-export * from './scrollToTop';
+export { getWidth } from './getWidth';
+export { scrollToTop } from './scrollToTop';
